Implement pg_autoctl create standby.
The monitor implements a basic role assignment policy that allows enforcing
that each group of servers has a primary and a standby. Beyond that, we will
need a way to either know how many standby servers are going to be deployed,
or a way to allow the operator to say that this new server is intended as an
extra standby.

This first patch in the PR allows the second case.