Use the synchronous_standby_names as computed by the monitor.
Also have the monitor implement the cases when we have a single node in the
group, or only two nodes in the group, or multiple standby nodes in the
group.

To be implemented correctly, pgautofailover.synchronous_standby_names() code
needs to depend on the current state of nodes. Then it's better to write the
code of this function in C and benefit from the existing functions that we
have that know how to deal with the state of the nodes in the monitor code.