Review replication slots handling in several cases.
In particular review the way creating a replication slot can be made
idempotent. It used to be that we would drop the slot and create it again.

In the case of multiple standby we might have failed to create one of
several slots, so we want to create-if-not-exists instead. That's what we do
now.

In passing, secure the dropping of "all" slots to be limited to physical
slots with a name that follows pg_auto_failover naming pattern. After all
other replication slots might be there that we don't own.