Reduce formation.number_sync_standbys when dropping standby nodes.
We need to ensure that number_sync_standbys + 1 <= standby_count to allow
for a good enough availability, in particular when using synchronous standby
servers. We already refused to raise number_sync_standbys when we don't have
enough standby nodes, and now we lower the value automatically when removing
the last standby that made it possible to reach it.