Improve and fix failover candidate selection mechanism.
We add more logging to be able to better debug production situations in that
complex process, and also fix some oddities. In particular the candidate
list was built separately twice where we now re-use explicitely the first
list we build.

This fixes a bug where we would select a candidate standby and try to
failover to a node that is currently in maintenance state.